refactor(foraging_cache): move impl modules to util/; split out postprocess
Move the implementation modules (parquet_builder, nwb_reader_aind,
nwb_reader_legacy) into a util/ sub-package, leaving build_cache and
query_examples as the top-level entry points. Split the per-subject coalescing
and triage-log helpers out of parquet_builder into util/postprocess.py
(coalesce_partitions, append_triage_log, read_csv/write_csv).

Remove the helpers orphaned by the global-discovery union refactor:
_enrich_with_co_assets, _add_s3_location_parallel, _get_assets_with_retry,
_parse_co_session_name, _read_existing_session_table, plus the now-unused
time / ThreadPoolExecutor imports. All imports updated; tests green.
