+++ tests/test_parquet_cache.py
@@ -18,7 +18,7 @@ import duckdb
 import pandas as pd
 import pyarrow.parquet as pq
 
-from aind_dynamic_foraging_data_utils.foraging_cache import parquet_builder
+from aind_dynamic_foraging_data_utils.foraging_cache.util import parquet_builder
 
 # ---------------------------------------------------------------------------
 # 1. _parse_nwb_filename
